环境部分代码与主函数分离
